[ALPS09851319] camsys: increase used_pipe bit
[Description]
increase used_pipe bit

[Test]
FCT

MTK-Commit-Id: 1075f3b4b1a393bba9afcd4cbd1a6195b53eb27b

CR-Id: ALPS09851319
Feature: [Android Default] Camera Application Basic Functions

+++ camsys/isp8s/cam/mtk_cam-request.h
@@ -109,7 +109,7 @@ struct mtk_cam_request {
 	bool is_buf_empty;
 
 	int used_ctx;
-	unsigned int used_pipe;
+	unsigned long used_pipe;
 
 	spinlock_t buf_lock;
 	struct list_head buf_list;
